+++ src/components/Footer/Footer.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react';
-import GitHub from '@/components/iconGitHub/GitHub/GitHub';
+import GitHub from '@/components/iconGitHub/iconGitHub';
 import RSLogo from './RSLogo';
 import styles from './footer.module.scss';
